requirements.txt
django
djangorestframework
djangorestframework-jwt
django-cors-headers
markdown
django-filter